renamed api.py to tra.py removed rest api calls from tra.py
